Jest testing environemtn set up, ready for proxy server
