+++ database/index.js
@@ -16,3 +16,5 @@ connection.connect(err => {
 });
 
 
+module.exports.connection = connection;
+
