+++ package.json
@@ -8,18 +8,28 @@
   },
   "scripts": {
     "react-dev": "webpack -d --watch",
-    "server-dev": "nodemon server/app.js",
-    "build": "webpack --config webpack.config.js"
+    "server-dev": "nodemon server/server.js",
+    "build": "webpack --config webpack.config.js",
+    "seed": "node database/seed.js",
+    "test": "jest"
   },
   "dependencies": {
-    "express": "^4.17.1"
+    "babel-jest": "^24.9.0",
+    "express": "^4.17.1",
+    "faker": "^4.1.0",
+    "mysql": "^2.17.1",
+    "superagent": "^5.1.0",
+    "supertest": "^4.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.6.2",
     "@babel/preset-env": "^7.6.2",
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.6",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.15.1",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor.git",
+    "jest": "^24.9.0",
     "nodemon": "^1.19.3",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
