+++ server/app.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const app = express();
-const port = 3008;
+// const port = 3008;
 app.use(express.static('client/dist'));
 
 const bodyParser = require('body-parser');
@@ -14,4 +14,6 @@ app.get('/', (req, res) => {
 
 
 
-app.listen(port, () => console.log(`Listening on port ${port}`));
+// app.listen(port, () => console.log(`Listening on port ${port}`));
+
+module.exports = app;
