+++ server/server.js
@@ -1,0 +1,4 @@
+const app = require('./app.js')
+const port = 3008;
+
+app.listen(port, () => console.log(`Listening on port ${port}`));
